refactor: use class-based seo manager and standardize titles
