+++ index.html
@@ -3,42 +3,42 @@
 	<head>
 		<meta charset="UTF-8" />
 		<link rel="icon" href="/favicon.ico" />
-                <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-                <meta name="description" content="Personal website of Gustavo Ocanto, engineering leader and software developer." />
-                <meta name="robots" content="index,follow" />
-                <meta name="theme-color" content="#ffffff" />
-                <link rel="canonical" href="https://example.com/" />
-                <title>Gustavo Ocanto</title>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta name="description" content="Personal website of Gustavo Ocanto, engineering leader and software developer." />
+		<meta name="robots" content="index,follow" />
+		<meta name="theme-color" content="#ffffff" />
+		<link rel="canonical" href="https://example.com/" />
+		<title>Home - Gustavo Ocanto</title>
 
-                <!-- Open Graph -->
-                <meta property="og:title" content="Gustavo Ocanto" />
-                <meta property="og:description" content="Personal website of Gustavo Ocanto, engineering leader and software developer." />
-                <meta property="og:type" content="website" />
-                <meta property="og:url" content="https://example.com/" />
-                <meta property="og:image" content="/favicon.ico" />
+		<!-- Open Graph -->
+		<meta property="og:title" content="Home - Gustavo Ocanto" />
+		<meta property="og:description" content="Personal website of Gustavo Ocanto, engineering leader and software developer." />
+		<meta property="og:type" content="website" />
+		<meta property="og:url" content="https://example.com/" />
+		<meta property="og:image" content="/favicon.ico" />
 
-                <!-- Twitter -->
-                <meta name="twitter:card" content="summary_large_image" />
-                <meta name="twitter:title" content="Gustavo Ocanto" />
-                <meta name="twitter:description" content="Personal website of Gustavo Ocanto, engineering leader and software developer." />
-                <meta name="twitter:image" content="/favicon.ico" />
+		<!-- Twitter -->
+		<meta name="twitter:card" content="summary_large_image" />
+		<meta name="twitter:title" content="Home - Gustavo Ocanto" />
+		<meta name="twitter:description" content="Personal website of Gustavo Ocanto, engineering leader and software developer." />
+		<meta name="twitter:image" content="/favicon.ico" />
 
-                <!-- Structured Data for AI and crawlers -->
-                <script type="application/ld+json">
-                        {
-                                "@context": "https://schema.org",
-                                "@type": "WebSite",
-                                "url": "https://example.com/",
-                                "name": "Gustavo Ocanto",
-                                "description": "Personal website of Gustavo Ocanto, engineering leader and software developer."
-                        }
-                </script>
+		<!-- Structured Data for AI and crawlers -->
+		<script type="application/ld+json">
+			{
+				"@context": "https://schema.org",
+				"@type": "WebSite",
+				"url": "https://example.com/",
+				"name": "Gustavo Ocanto",
+				"description": "Personal website of Gustavo Ocanto, engineering leader and software developer."
+			}
+		</script>
 
-                <script defer>
-                        if (localStorage.getItem('dark-mode') === 'true') {
-                                document.documentElement.classList.add('dark');
-                        }
-                </script>
+		<script defer>
+			if (localStorage.getItem('dark-mode') === 'true') {
+				document.documentElement.classList.add('dark');
+			}
+		</script>
 	</head>
 	<body class="font-sans antialiased bg-white text-slate-800 dark:bg-slate-900 dark:text-slate-200 tracking-tight">
 		<div id="app"></div>
